Add SceneGrid class / Optimise scene view and FPS rate
