--- CadEditor/MeshObjects/IMeshEnumerable.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CadEditor.MeshObjects
-{
-    public interface IMeshEnumerable
-    {
-        //Object3D Contains(Object3D object3D);
-
-        //List<Object3D> Clone();
-
-
-    }
-}
