+++ CadEditor/Scene/ExtraObjects/SceneCollection.cs
