+++ CadEditor/Scene/ISceneObject.cs
@@ -1,14 +1,18 @@
 ﻿
 namespace CadEditor
 {
-    public interface ISceneObject
+    public interface IDrawable
+    {
+        void Draw();
+    }
+
+    public interface ISceneObject: IDrawable
     {
         ISceneObject ParentObject { get; set; }
         bool IsSelected { get; set; }
 
         Point3D GetCenterPoint();
         void Move(Vector v);
-        void Draw();
         void Select();
         void Deselect();
         ISceneObject CheckSelected(Ray ray);
